// src/candidates/_dtos/update-candidate.dto.ts
export interface UpdateCandidateDto {
    name?: string;
    lastName?: string;
    dni?: string;
    email?: string;
    phone?: string;
    isAvailable?: boolean;
    city?: string;
    technology?: string;
    experience?: Date;
    isRemote?: boolean;
    photo?: string;
    comments?: string;
    linkedIn?: string;
    noticePeriod?: number;
}